requirements.txt
Flask==3.0.3
pymongo==4.8.0b0
PyJWT==2.4.0
validator===0.7.1